Separated PropertyCache from existing module
