+++ src/main/java/fr/estecka/variantscit/modules/AMultiComponentCachingModule.java
@@ -1,15 +1,11 @@
 package fr.estecka.variantscit.modules;
 
-import java.lang.ref.WeakReference;
-import java.lang.ref.ReferenceQueue;
 import java.util.stream.Stream;
 import org.jetbrains.annotations.Nullable;
-import fr.estecka.variantscit.VariantsCitMod;
+import fr.estecka.variantscit.MultiPropertyCache;
 import fr.estecka.variantscit.api.ICitModule;
 import fr.estecka.variantscit.api.IVariantManager;
 import fr.estecka.variantscit.format.properties.IStringProperty;
-import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
-import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
 
@@ -30,14 +26,11 @@ abstract class AMultiComponentCachingModule
 implements ICitModule
 {
 	protected final boolean debug;
-	private final IStringProperty[] properties;
-
-	private final Int2ObjectMap<CacheEntry> hashToVariant = new Int2ObjectOpenHashMap<>();
-	private final ReferenceQueue<Object> expiredComponents = new ReferenceQueue<>();
+	private final MultiPropertyCache cache;
 
 	protected AMultiComponentCachingModule(boolean debug, Stream<IStringProperty> properties){
 		this.debug = debug;
-		this.properties = properties.distinct().toArray(IStringProperty[]::new);
+		this.cache = new MultiPropertyCache(properties, debug);
 	}
 
 	// TODO: Ensure child classes can't access unregistered non-cached components.
@@ -45,76 +38,6 @@ implements ICitModule
 
 	@Override
 	public final Identifier GetItemModel(ItemStack stack, IVariantManager library){
-		this.ExpungeExpiredEntries();
-
-		int hash = this.HashStack(stack);
-		CacheEntry entry = this.hashToVariant.get(hash);
-		if (entry == null) {
-			entry = this.CreateEntry(hash, stack, library);
-			if (debug)
-				VariantsCitMod.LOGGER.info("[AMultiComponent] Cache size: {}; Model Id: {}", hashToVariant.size(), String.valueOf(entry.variant));
-		}
-		return entry.variant;
+		return this.cache.ComputeIfAbsent(stack, s->this.RecomputeItemModel(s, library));
 	}
-
-
-	/**
-	 * @see {@linkplain java.util.Arrays#hashCode(Object[])}
-	 */
-	private int HashStack(ItemStack stack){
-		int hash = 17;
-		for (var prop : this.properties){
-			hash = hash*31 + prop.GetPropertyHash(stack);
-		}
-		return hash;
-	}
-
-	/**
-	 * TODO: As-is, an entry where all registered components are null will never
-	 * expire. This is limited to one entry per module, so it is negligible.
-	 */
-	private CacheEntry CreateEntry(int hash, ItemStack stack, IVariantManager library){
-		Identifier variant = this.RecomputeItemModel(stack, library);
-		WeakReference<?>[] weakRefs = new WeakReference[properties.length];
-
-		for (int i=0; i<properties.length; ++i){
-			Object cmp = properties[i].GetReference(stack);
-			if (cmp != null)
-				weakRefs[i] = new HashedWeakReference(hash, cmp, this.expiredComponents);
-		}
-
-		CacheEntry entry = new CacheEntry(variant, weakRefs);
-		this.hashToVariant.put(hash, entry);
-		return entry;
-	}
-
-	private void ExpungeExpiredEntries(){
-		HashedWeakReference weakRef;
-		while ((weakRef=(HashedWeakReference)expiredComponents.poll()) != null){
-			this.hashToVariant.remove(weakRef.hash);
-		}
-	}
-
-	static private class HashedWeakReference
-	extends WeakReference<Object>
-	{
-		/**
-		 * The key of the associated entry that must be cleared along with this
-		 * reference.
-		 */
-		public final int hash;
-
-		public HashedWeakReference(int hash, Object referent, ReferenceQueue<Object> queue){
-			super(referent, queue);
-			this.hash = hash;
-		}
-	}
-
-	/**
-	 * Weak references are kept around so that the weak reference itself doesn't
-	 * get garbage collected  before its referee. Otherwise, references will not
-	 * get enqueued, and the cache will never be cleared.
-	 */
-	static private record CacheEntry(Identifier variant, WeakReference<?>[] components)
-	{}
 }
